Pass release ids instead of instances
The information contained in a `discogs_client.models.Release`
object depends on how the instance was created. Often, only a
subset of the full release data is retrieved from Discogs, and the
formatting of the `title` property is inconsistent (sometimes
'title', sometimes 'artist - title'). If the instance is requested
directly via it's release id, the full set of data is retrieved.

Therefore, despite the resultant increase in API calls, it's worth
passing around only the ids of search results, then re-requesting
the release when more information is needed. This should ensure
that data are consistent when later formatted.

While it would be possible to simply refresh the incomplete
instance, this way allows simpler storage of state and seperation
of processess, since only an id is needed to resume operation.
Later, if the rate-limiting becomes an issue, the code can be
refactored to take both ids and release instances, refreshing only
if data is missing, although the saving will be minor.