Move nofuture script logic to __main__
`nofuture` was a hybrid of a module for decompression wrappers and
a script to run the first stage of the system (decompression).
It should be reduced to simply the decompression functions, so the
script logic has been moved to the new __main__ script.

Although they're only used in the stage after decompression, the
release path formatting functions remain here until they're later
moved to somewhere more appropriate.